refactor: moved new options pkg to libsq/core/options
