+++ cli/cli.go
@@ -24,9 +24,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/neilotoole/sq/cli/config/options"
+	"github.com/neilotoole/sq/libsq/core/options"
 
-	"github.com/neilotoole/sq/cli/config/options/format"
+	"github.com/neilotoole/sq/cli/output/format"
 
 	"github.com/neilotoole/sq/cli/flag"
