+++ cli/complete.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/neilotoole/sq/cli/config/options"
+	"github.com/neilotoole/sq/libsq/core/options"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
