+++ cli/output/jsonw/configw.go
@@ -3,7 +3,7 @@ package jsonw
 import (
 	"io"
 
-	"github.com/neilotoole/sq/cli/config/options"
+	"github.com/neilotoole/sq/libsq/core/options"
 
 	"github.com/neilotoole/sq/cli/output"
 )
