+++ drivers/csv/import.go
@@ -18,9 +18,9 @@ import (
 
 	"github.com/neilotoole/sq/libsq/core/lg"
 
-	options2 "github.com/neilotoole/sq/cli/config/options"
 	"github.com/neilotoole/sq/libsq"
 	"github.com/neilotoole/sq/libsq/core/errz"
+	"github.com/neilotoole/sq/libsq/core/options"
 	"github.com/neilotoole/sq/libsq/driver"
 	"github.com/neilotoole/sq/libsq/source"
 )
@@ -28,14 +28,14 @@ import (
 // OptImportHeader allows the user to specify whether the imported CSV
 // file has a header or not. If not set, the importer will try to
 // detect if the input has a header.
-var OptImportHeader = options2.NewBool("driver.csv.header", false, "")
+var OptImportHeader = options.NewBool("driver.csv.header", false, "")
 
 // OptEmptyAsNull determines if an empty CSV field is treated as NULL
 // or as the zero value for the kind of that field.
-var OptEmptyAsNull = options2.NewBool("driver.csv.empty-as-null", true, "")
+var OptEmptyAsNull = options.NewBool("driver.csv.empty-as-null", true, "")
 
 // OptDelim specifies the CSV delimiter to use.
-var OptDelim = options2.NewString(
+var OptDelim = options.NewString(
 	"driver.csv.delim",
 	"comma",
 	`Possible values are: comma, space, pipe, tab, colon, semi, period.
@@ -188,7 +188,7 @@ func getDelimiter(src *source.Source) (rune, error) {
 // getDelimFromOptions returns ok as true and the delimiter rune if a
 // valid value is provided in src.Options, returns ok as false if
 // no valid value provided, and an error if the provided value is invalid.
-func getDelimFromOptions(opts options2.Options) (r rune, ok bool, err error) {
+func getDelimFromOptions(opts options.Options) (r rune, ok bool, err error) {
 	if len(opts) == 0 {
 		return 0, false, nil
 	}
