+++ libsq/core/options/testdata/good.01.yml
