+++ libsq/driver/driver.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/neilotoole/sq/cli/config/options"
+	"github.com/neilotoole/sq/libsq/core/options"
 
 	"github.com/neilotoole/sq/libsq/driver/dialect"
